pyproject.toml
[tool.poetry]
name = "test-task-inforce"
version = "0.1.0"
description = ""
authors = ["Victor <polyakov.victor21@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
